backend/requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
alabaster==0.7.12
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.13.4
blinker==1.4
certifi==2020.6.20
cffi==1.17.1
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.7
cloud-init==24.4
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
cryptography==44.0.2
dataclasses-json==0.6.7
datasets==3.3.2
dateparser==1.2.1
dbus-python==1.2.18
dill==0.3.8
diskcache==5.6.3
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
docutils==0.17.1
exceptiongroup==1.2.2
eyantra-autoeval==0.1.47
fastapi==0.115.11
filelock==3.6.0
frozenlist==1.5.0
fsspec==2024.12.0
git-filter-repo==2.47.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.2
hyperlink==21.0.0
identify==2.4.10
idna==3.3
imagesize==1.3.0
importlib-metadata==4.6.4
incremental==21.3.0
invoice2data==0.4.5
jeepney==0.7.1
Jinja2==3.0.3
jiter==0.8.2
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.43
langchain-text-splitters==0.3.6
langsmith==0.3.13
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llama_cpp_python==0.3.7
MarkupSafe==2.0.1
marshmallow==3.26.1
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.4.2
nodeenv==0.13.4
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
openai==1.65.4
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pdfminer.six==20240706
Pillow==9.0.1
platformdirs==2.5.1
pre-commit==2.17.0
propcache==0.3.0
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.22
pycurl==7.44.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymongo==4.11.3
pyOpenSSL==25.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==5.4.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==12.6.0
roman==3.3
safetensors==0.5.3
scipy==1.15.2
SecretStorage==3.3.1
sentencepiece==0.2.0
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.7
Sphinx==4.3.2
SQLAlchemy==2.0.38
starlette==0.46.0
sympy==1.13.1
systemd-python==234
tenacity==9.0.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
Twisted==22.1.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.34.0
virtualenv==20.13.0+ds
wadllib==1.3.6
xgboost==3.0.0
xxhash==3.5.0
yarl==1.18.3
zipp==1.0.0
zope.interface==5.4.0
zstandard==0.23.0
